refactor: stub 헬퍼 추가 및 서비스 테스트 분리
