+++ src/test/java/roomescape/application/ReservationQueryServiceTest.java
@@ -19,7 +19,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import roomescape.application.dto.ReservationCreateDto;
 import roomescape.application.dto.ReservationDto;
 import roomescape.application.dto.ReservationWaitingDto;
-import roomescape.application.dto.ThemeDto;
 import roomescape.domain.Member;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationStatus;
@@ -27,29 +26,24 @@ import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
 import roomescape.domain.Waiting;
 import roomescape.domain.repository.ReservationRepository;
+import roomescape.testFixture.StubHelper;
 
 @ExtendWith(MockitoExtension.class)
-public class ReservationServiceTest {
+public class ReservationQueryServiceTest {
 
     @InjectMocks
-    private ReservationService reservationService;
+    private ReservationQueryService reservationService;
     @Mock
     private ReservationRepository reservationRepository;
     @Mock
-    private TimeService timeService;
-    @Mock
-    private ThemeService themeService;
-    @Mock
     private MemberService memberService;
-    @Mock
-    private ClockProvider clockProvider;
 
     @DisplayName("사용자 예약 목록 조회 성공")
     @Test
     void getReservationsByMember() {
         //given
         Long memberId = 2L;
-        Member member = stubMember(memberId);
+        Member member = StubHelper.stubMember(memberId, memberService);
 
         Theme theme = Theme.of(1L, "테마1", "테마1입니다.", "썸네일1");
         ReservationTime time = ReservationTime.of(1L, LocalTime.of(10, 0));
@@ -73,48 +67,5 @@ public class ReservationServiceTest {
                 () -> Assertions.assertThat(dtos.get(0).reservationId()).isEqualTo(reservation1.getId()),
                 () -> Assertions.assertThat(dtos.get(1).reservationId()).isEqualTo(reservation2.getId())
         );
-    }
-
-    @DisplayName("사용자 예약 추가")
-    @Test
-    void registerReservation() {
-        //given
-        Theme theme = stubTheme(1L);
-        ReservationTime time = stubTime(1L);
-        Member member = stubMember(1L);
-        Waiting waiting = Mockito.mock(Waiting.class);
-
-        Reservation reservation = Reservation.of(1L, member, theme, LocalDate.now().plusDays(1), time, waiting);
-        Mockito.doReturn(reservation).when(reservationRepository).save(Mockito.any(Reservation.class));
-        Mockito.doReturn(LocalDateTime.now()).when(clockProvider).now();
-
-        //when
-        ReservationDto reservationDto = reservationService.registerReservation(new ReservationCreateDto(
-                reservation.getDate(),
-                theme.getId(),
-                time.getId(),
-                member.getId()
-        ));
-
-        //then
-        Assertions.assertThat(reservationDto).isEqualTo(ReservationDto.from(reservation));
-    }
-
-    private Theme stubTheme(long themeId) {
-        Theme theme = Theme.of(themeId, "테마1", "테마1입니다.", "썸네일1");
-        Mockito.doReturn(theme).when(themeService).getThemeById(themeId);
-        return theme;
-    }
-
-    private ReservationTime stubTime(long timeId) {
-        ReservationTime time = ReservationTime.of(timeId, LocalTime.of(10, 0));
-        Mockito.doReturn(time).when(timeService).getTimeEntityById(timeId);
-        return time;
-    }
-
-    private Member stubMember(long memberId) {
-        Member member = Member.of(memberId, "브라운", "brown@email.com", "brown", USER);
-        Mockito.doReturn(member).when(memberService).getMemberEntityById(memberId);
-        return member;
     }
 }
